AUDIT-28: Move backfill into Liquibase CustomTaskChange
