+++ api/src/main/resources/org/openmrs/liquibase/updates/liquibase-update-to-latest-3.0.x.xml
@@ -141,6 +141,12 @@
 			<column name="description" type="varchar(255)"/>
 		</addColumn>
 	</changeSet>
-	
+
+	<changeSet id="AUDIT-28-2026-03-19" author="Binayak490-cyber">
+		<comment>Backfill pre-existing rows into Envers audit tables so that the audit UI can resolve
+			references to entities that existed before auditing was enabled</comment>
+		<customChange class="org.openmrs.util.databasechange.BackfillEnversAuditTablesChangeset"/>
+	</changeSet>
+
 </databaseChangeLog>
